+++ src/index.tsx
@@ -3,14 +3,20 @@ import ReactDOM from 'react-dom';
 import * as Sentry from '@sentry/browser';
 
 import './index.css';
+import packageJson from '../package.json';
+import configService from './config/configService';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
+// Checks that required configs were provided during build. This makes
+// badly configured environments easier to detect.
+configService.checkConfigs();
+
 if (process.env.NODE_ENV !== 'development') {
   Sentry.init({
-    dsn: process.env.REACT_APP_SENTRY_DSN,
-    environment: process.env.REACT_APP_ENVIRONMENT,
-    release: `${process.env.REACT_APP_APPLICATION_NAME}@${process.env.REACT_APP_VERSION}`,
+    dsn: configService.getConfig('REACT_APP_SENTRY_DSN'),
+    environment: configService.getConfig('REACT_APP_ENVIRONMENT'),
+    release: `${packageJson.name}@${packageJson.version}`,
   });
 }