+++ CHANGELOG.md
@@ -3,6 +3,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## Unreleased
+
+## [1.3.0] - 2020-02-18
+
+### Changed
+- [#77](https://github.com/City-of-Helsinki/youth-membership-admin-ui/pull/77) Use GitHub actions instead of Travis
+- [#76](https://github.com/City-of-Helsinki/youth-membership-admin-ui/pull/76) Use federated youth membership service
+- [#84](https://github.com/City-of-Helsinki/youth-membership-admin-ui/pull/84) Use public address for Jassari API
+
+### Fixed
+- [#92](https://github.com/City-of-Helsinki/youth-membership-admin-ui/pull/92) Non-required youth profile fields can be cleared
+
+## [1.2.2] - 2020-02-01
+- Re-release on v.1.2.0 with backend pointing to profile.
+
+## [1.2.1] - 2020-02-01
+- Re-release on v.1.2.0 with production GitHub actions workflow.
+
 ## [1.2.0] - 2020-11-03
 ### Added
 - E2E tests added for registering, searching, viewing and modifying youths